Add --force-hwaccel flag
Hw accellerated decode seems to slow things down.  Decoding is faster,
but transfer of data to the CPU apparently is slower than CPU decoding.

This is very fast:
 ffmpeg -hwaccel vaapi -hwaccel_output_format vaapi -i x.mp4 -f null -

This is slow:
 ffmpeg -hwaccel vaapi -hwaccel_output_format vaapi -i x.mp4 \
 -vf "hwdownload,format=nv12,format=yuv420p" -f rawvideo -y /dev/null

Even slower than:
 ffmpeg -i x.mp4 -vf "format=yuv420p" -f rawvideo -y /dev/null

So unless a better way of transferring data to the CPU can be found,
hw decode should not be the default.